pyproject.toml
[project]
name = "cpp-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["cairosvg>=2.8.2", "pillow>=11.3.0"]
